Changes Made to Room for Change
Uhh I'm not quite sure if the other stuff was intentionally made that way but here are the stuff that I've changed or added so far:

* Upon getting a heart (which adds / health++) that is more that the maxHealth, it won't be added anymore and will stay as health = maxHealth.
* Added a PauseScreen, if the Key.P or P is pressed while on the GameScreen, will change or Pause the game and sounds. If pressed again will Resume the game and sounds. (since it's hard if it doesn't get paused >.> everything updates monsters also might run over you).
* Changed / edited the MainMenuScreen's Background, since I added a key or the pause feature by pressing P. I added the letter command at the MainMenu. (quite hard since I don't have the Font for it so I manually did it).

Ohh I added an icon for the desktop since it would be cool if it has lol. (although it's just a shrinked version of the background :/)

It was fun fiddling with it! and I learned a lot !